+++ packages/jivs-engine/tests/Conditions/FluentValidatorCollectorExtensions.test.ts
@@ -6,11 +6,15 @@ import { ValidatorConfig } from '../../src/Interfaces/Validator';
 import {
     AllMatchConditionConfig, AnyMatchConditionConfig, CountMatchesConditionConfig, DataTypeCheckConditionConfig,
     EqualToConditionConfig, EqualToValueConditionConfig, GreaterThanConditionConfig, GreaterThanOrEqualConditionConfig, GreaterThanOrEqualValueConditionConfig, GreaterThanValueConditionConfig, IntegerConditionConfig, LessThanConditionConfig,
-    LessThanOrEqualConditionConfig, LessThanOrEqualValueConditionConfig, LessThanValueConditionConfig, MaxDecimalsConditionConfig, NotConditionConfig, NotEqualToConditionConfig, NotEqualToValueConditionConfig, NotNullConditionConfig, PositiveConditionConfig, RangeConditionConfig,
+    LessThanOrEqualConditionConfig, LessThanOrEqualValueConditionConfig, LessThanValueConditionConfig, MaxDecimalsConditionConfig, NotEqualToConditionConfig, NotEqualToValueConditionConfig, NotNullConditionConfig, PositiveConditionConfig, RangeConditionConfig,
     RegExpConditionConfig, RequireTextConditionConfig, StringLengthConditionConfig
 } from '../../src/Conditions/ConcreteConditions';
+import {
+    NotConditionConfig
+} from '../../src/Conditions/NotCondition';
 import { ConditionEvaluateResult } from '../../src/Interfaces/Conditions';
 import { MockValidationServices } from '../TestSupport/mocks';
+import { WhenConditionConfig } from '../../src/Conditions/WhenCondition';
 
 function createFluent(): ValidationManagerStartFluent {
     return new ValidationManagerStartFluent(null, new MockValidationServices(true, true));
@@ -2031,10 +2035,129 @@ describe('not with ValidationManagerStartFluent', () => {
     });    
     test('Null as the function parameter throws', () => {
         let fluent = createFluent();
-        expect(()=> fluent.input('Field1').not(null!, 'Error')).toThrow(/condition/);
+        expect(()=> fluent.input('Field1').not(null!, 'Error')).toThrow(/childConfig/);
     });
     test('Non-function as the function parameter throws', () => {
         let fluent = createFluent();
         expect(() => fluent.input('Field1').not({} as any, 'Error')).toThrow(/Function expected/);
     });    
+});
+
+describe('when with ValidationManagerStartFluent', () => {
+    test('With empty condition in both enabler and childCondition, creates ValidatorConfig with WhenCondition with type=When, enablerConfig={} and childConditionConfig={}', () => {
+
+        let testItem = createFluent().input('Field1').when(
+            (enabler) => enabler,
+            (child) => child);
+        TestFluentValidatorCollector(testItem, <ValidatorConfig>{
+            conditionConfig: <WhenConditionConfig>{
+                conditionType: ConditionType.When,
+                enablerConfig: {},
+                childConditionConfig: {}
+            }
+        });
+    });
+    test('With child condition setup with requireText and enabler condition with regexp, creates ValidatorConfig with WhenCondition with type=When and both enablerConfig and childConditionConfigs populated', () => {
+
+        let testItem = createFluent().input('Field1')
+            .when((enabler)=> enabler.regExp(/abc/),
+                (child) => child.requireText(null, 'F1'));
+        TestFluentValidatorCollector(testItem, <ValidatorConfig>{
+            conditionConfig: <WhenConditionConfig>{
+                conditionType: ConditionType.When,
+                enablerConfig: <any>{
+                    conditionType: ConditionType.RegExp,
+                    expression: /abc/
+                },
+                childConditionConfig: <any>{
+                    conditionType: ConditionType.RequireText,
+                    valueHostName: 'F1'
+                }
+            }
+        });
+    });
+    test('With errorMessage and parameter.summaryMessage creates ValidatorConfig with WhenCondition with only type assigned and errorMessage + summaryMessage assigned', () => {
+
+        let testItem = createFluent().input('Field1')
+            .when(
+                (enabler)=> enabler.regExp(/abc/, null, null, 'F2'),
+                (child) => child.requireText(null, 'F1'),
+                
+                'Error', { summaryMessage: 'Summary' });
+        TestFluentValidatorCollector(testItem, <ValidatorConfig>{
+            conditionConfig: <WhenConditionConfig>{
+                conditionType: ConditionType.When,
+                enablerConfig: <any>{
+                    conditionType: ConditionType.RegExp,
+                    expression: /abc/,
+                    valueHostName: 'F2'
+                },
+                childConditionConfig: <any>{
+                    conditionType: ConditionType.RequireText,
+                    valueHostName: 'F1'
+                }
+            },
+            errorMessage: 'Error',
+            summaryMessage: 'Summary'
+        });
+    });
+    test('With errorMessage = null, parameter.errorMessage and parameter.summaryMessage creates ValidatorConfig with WhenCondition with only type assigned and errorMessage + summaryMessage assigned', () => {
+
+        let testItem = createFluent()
+            .input('Field1').when(
+                (enabler)=> enabler,    
+                (child) => child, null, { errorMessage: 'Error', summaryMessage: 'Summary' });
+        TestFluentValidatorCollector(testItem, <ValidatorConfig>{
+            conditionConfig: <WhenConditionConfig>{
+                conditionType: ConditionType.When,
+                enablerConfig: {},
+                childConditionConfig: {}
+            },
+            errorMessage: 'Error',
+            summaryMessage: 'Summary'
+        });
+    });
+    test('With errorMessage assigned, parameter.errorMessage and parameter.summaryMessage creates ValidatorConfig with WhenCondition with only type assigned. ErrorMessage is from first parameter, when validatorConfig assigned', () => {
+
+        let testItem = createFluent().input('Field1').when(
+            (enabler) => enabler,
+            (child) => child,
+            'FirstError', { errorMessage: 'SecondError' });
+        TestFluentValidatorCollector(testItem, <ValidatorConfig>{
+            conditionConfig: <WhenConditionConfig>{
+                conditionType: ConditionType.When,
+                enablerConfig: {},  
+                childConditionConfig: {}
+            },
+            errorMessage: 'FirstError'
+        });
+    });
+    test('When there are 2 child conditions, throws', () => {
+        expect(() => createFluent().input('Field1')
+            .when((enabler)=> enabler,
+                (child) => child.requireText(null, 'F1').requireText(null, 'F2'))).toThrow();
+    });    
+    test('When there are 2 enabler conditions, throws', () => {
+        expect(() => createFluent().input('Field1')
+            .when((enabler)=> enabler.requireText(null, 'F1').requireText(null, 'F2'),
+                (child) => child)).toThrow();
+    });        
+    test('Null as the enabler function parameter throws', () => {
+        let fluent = createFluent();
+        expect(() => fluent.input('Field1').when(null!, (child) => child,
+            'Error')).toThrow(/enabler/);
+    });
+    test('Null as the child condition function parameter throws', () => {
+        let fluent = createFluent();
+        expect(() => fluent.input('Field1').when((enabler)=> enabler, null!,
+            'Error')).toThrow(/childConfig/);
+    });    
+    test('Non-function as the child function parameter throws', () => {
+        let fluent = createFluent();
+        expect(() => fluent.input('Field1').when((enabler)=> enabler, {} as any, 'Error')).toThrow(/Function expected/);
+    });    
+    test('Non-function as the enabler function parameter throws', () => {
+        let fluent = createFluent();
+        expect(() => fluent.input('Field1').when({} as any, (child)=>child, 'Error')).toThrow(/Function expected/);
+    });        
 });