+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Extensive examples for configuration are now in jivs-examples.
 - **Breaking API change** Fluent syntax for all, any, and countMatches revised to take a function instead of a FluentConditionCollector
 - New condition, NotCondition. Inverts the result of a child condition. input('field').not((child)=> child.regExp(/\d*/));
+- New condition, WhenCondition. Replacement for the Enabler feature on ValidatorConfig. Has two conditions: enabler and the condition to enable. The child condition is executed only when the enabler condition matches the condition. input('field').when((enabler)=>enabler.regExp(/\d*/, null, null, 'otherValueHostName'), (child)=>child.requireText())
 ## 0.17.1
 - Added setIssuesFound() method on ValidationManager and ValidatableValueHostBase. Simplifies how to send Jivs errors found by the server and sent up to the client.
 - Added {DataType} token to error messages, with localization using TextLocalizerService.